Merge pull request #208 from novopl/release/0.24

+++ src/peltak/logic/docs.py
@@ -23,12 +23,14 @@ import sys
 
 # local imports
 from peltak.core import conf
+from peltak.core import context
 from peltak.core import log
 from peltak.core import shell
+from peltak.core import util
 
 
-def docs(recreate, gen_index, run_doctests, verbose):
-    # type: (bool, bool, bool, int) -> None
+def docs(recreate, gen_index, run_doctests):
+    # type: (bool, bool, bool) -> None
     """ Build the documentation for the project.
 
     Args:
@@ -41,8 +43,9 @@ def docs(recreate, gen_index, run_doctests, verbose):
         run_doctests (bool):
             Set to **True** if you want to run doctests after the documentation
             is generated.
-        verbose (int):
-            The verbosity level.
+        pretend (bool):
+            If set to **True**, do not actually execute any shell commands, just
+            print the command that would be executed.
     """
     build_dir = conf.get_path('build_dir', '.build')
     docs_dir = conf.get_path('docs.path', 'docs')
@@ -59,7 +62,7 @@ def docs(recreate, gen_index, run_doctests, verbose):
                 shutil.rmtree(path)
 
     if refdoc_paths:
-        gen_ref_docs(verbose, gen_index)
+        gen_ref_docs(gen_index)
     else:
         log.err('Not generating any reference documentation - '
                 'No docs.reference specified in config')
@@ -85,7 +88,7 @@ def docs(recreate, gen_index, run_doctests, verbose):
         ))
 
 
-def gen_ref_docs(verbose, gen_index=False):
+def gen_ref_docs(gen_index=False):
     # type: (int, bool) -> None
     """ Generate reference documentation for the project.
 
@@ -93,9 +96,6 @@ def gen_ref_docs(verbose, gen_index=False):
     reference documentation.
 
     Args:
-
-        verbose (int):
-            Verbosity level. This will be passed directly to sphinx-refdoc.
         gen_index (bool):
             Set it to **True** if you want to generate the index file with the
             list of top-level packages. This is set to default as in most cases
@@ -113,18 +113,22 @@ def gen_ref_docs(verbose, gen_index=False):
         log.err("Exception: {}".format(ex))
         sys.exit(-1)
 
+    pretend = context.get('pretend', False)
+
     docs_dir = conf.get_path('docs.path', 'docs')
     docs_ref_dir = os.path.join(docs_dir, 'ref')
     refdoc_paths = conf.get('docs.reference', [])
 
-    log.info('Removing previously generated reference documentation')
     if os.path.exists(docs_ref_dir):
-        shutil.rmtree(docs_ref_dir)
+        if not pretend:
+            log.info('Removing existing reference docs')
+            shutil.rmtree(docs_ref_dir)
+        else:
+            log.info('Would remove old reference docs')
 
-    log.info('Generating reference documentation')
     args = {
         'out_dir': docs_ref_dir,
-        'verbose': verbose,
+        'verbose': context.get('verbose', 0),
     }
 
     if gen_index:
@@ -132,4 +136,10 @@ def gen_ref_docs(verbose, gen_index=False):
 
     pkg_paths = [conf.proj_path(p) for p in refdoc_paths]
 
-    generate_docs(pkg_paths, **args)
+    if not pretend:
+        log.info('Generating reference documentation')
+        generate_docs(pkg_paths, **args)
+    else:
+        log.info("Would generate reference docs with the following params")
+        shell.cprint('<90>{}', util.yaml_dump(args).rstrip())
+        shell.cprint('<90>paths:\n<34>{}', util.yaml_dump(pkg_paths).rstrip())
